Move validation.ts to utils/is-item-valid.ts
