+++ packages/dataviews/src/stories/dataform.story.tsx
@@ -17,7 +17,7 @@ import {
  * Internal dependencies
  */
 import DataForm from '../components/dataform';
-import { isItemValid } from '../validation';
+import isItemValid from '../utils/is-item-valid';
 import type {
 	Field,
 	Form,
